Passport_check.py: refactored out passport validity checking
