+++ AoC_2020_Day_4/test/test_day_4_1.py
@@ -1,10 +1,5 @@
 from src.Day_4 import Day_4 as D4
-
-def valid(mandatory_fields, line):
-    for mandatory_field in mandatory_fields:
-        if mandatory_field not in line:
-            return False
-    return True
+from Passport_check import Passport_check as PC
 
 def test_example_config():
     file = 'data/aoc_2020_day_4_data_example.txt'
@@ -18,4 +13,4 @@ def get_nbr_of_valid_passports(file):
     with open(file) as f:
         lines = [line.strip() for line in f]
 
-    return D4.get_nbr_of_valid_passports(valid, lines)
+    return D4.get_nbr_of_valid_passports(PC.valid, lines)
